+++ lua/plugins/editing.lua
@@ -11,28 +11,38 @@ return {
     opts = {},
   },
 
-  -- commenting with e.g. `gcc` or `gcip`
-  -- respects TS, so it works in quarto documents
-  {
+  { -- commenting with e.g. `gcc` or `gcip`
+    -- respects TS, so it works in quarto documents 'numToStr/Comment.nvim',
     'numToStr/Comment.nvim',
     version = nil,
+    cond = function()
+      return vim.fn.has 'nvim-0.10' == 0
+    end,
     branch = 'master',
     config = true,
+  },
+
+  { -- format things as tables
+    'godlygeek/tabular',
   },
 
   { -- Autoformat
     'stevearc/conform.nvim',
     enabled = true,
+    keys = {
+      { '<leader>cf', '<cmd>lua require("conform").format()<cr>', desc = "[f]ormat" },
+    },
     config = function()
       require('conform').setup {
         notify_on_error = false,
-        format_on_save = {
-          timeout_ms = 500,
-          lsp_fallback = true,
-        },
+        -- format_on_save = {
+        --   timeout_ms = 500,
+        --   lsp_fallback = true,
+        -- },
         formatters_by_ft = {
           lua = { 'mystylua' },
           python = { 'isort', 'black' },
+          quarto = { 'injected' },
         },
         formatters = {
           mystylua = {
@@ -82,6 +92,7 @@ return {
 
   {
     'chrishrb/gx.nvim',
+    enabled = false,
     keys = { { 'gx', '<cmd>Browse<cr>', mode = { 'n', 'x' } } },
     cmd = { 'Browse' },
     init = function()